requirements.txt
Flask==2.3.2
mysql.connector
docxtpl
datetime
flask_cors
python-dotenv